packages/backend-server/requirements.txt
fastapi
gunicorn
pymongo
structlog
uvicorn
redis
pydantic
pydantic-settings
firebase-admin
pycryptodome
pytz
jinja2
logfire[fastapi]